Add sampledata app that loads test data from other apps
Provides a load_sampledata mangement command that, when run, will look
through all registered apps in the project and attempts to import a
sampledata module.

In this module, applications should fill the database with test data. We
will use this to replace the current JSON fixtures with the model makers
from the tests.