106. Start work on suggestions for videos to be added
This will allow site administrators to collect suggestions for videos or
conferences to be added to the site. A suggestion has different
states, it's either new, in progress, completed, rejected or marked as
spam. When completing or rejected a suggestion, the site owner may add a
few words about the resolution.

This commit establishes a base to work from, that is a database table,
integration into the admin and a simple list of all suggestions with
their status for users to see. Suggestions marked as spam will not be
shown.