119. Add support for sitemap.xml
Django already ships with a contrib app that implements the sitemap
protocol, so this change is straightforward.

This adds sitemaps for pages, categories, speakers and videos. Tags
don't have their own page yet, so we'll add a sitemap for them later.

Because we're using Jinja2 for templates, the sitemaps application has
to be ignored by jingo, otherwise rendering will fail due to unknown
tags. Typically, jingo determines the app's name by looking at the
template name, the sitemaps application however just passes
'sitemap.xml' as template name. Therefore we need to use that as
application name in JINGO_EXCLUDE_APPS.