API overhaul
This fixes a lot of validation issues when posting videos and adds tests
for a lot of them, too.

This also fixes the problem where if you do a GET or a DELETE on a Video
with an api_key, you'd get back an HTTP 500 saying that there were two
things with that pk (which is pretty crazy).