Fix problems with POST empty string title with api
If you try to POST a video with an empty string title, then it'd create
it and the slug would be based on the empty string. That's total ick
since all videos should require a title.

This fixes that by requiring a title. Also adds a test for that
situation.

Also adds a big comment about data checking in hydrate.