Fix API slug/title/tags/speakers issues
This tweaks the code so that it's saving the speakers and tags data
in the correct place in the tastypie lifecycle. This fixes the problem
where we were saving the Video in hydrate in order to add speakers and
tags to it, but that was before it was fully hydrated which caused
problems with empty titles and slugs.