Issue #157. Auto login without BrowserID
This is useful when working on richard when authenticating with
BrowserID/Persona is not possible, e.g. when there is no internet
connection.

To enable the auto login, you have to set `BROWSERID_AUTOLOGIN` to the
email address of the user you want to login as.

When enabled, the django-browserid JavaScript will be replaced with a
mock that calls the server-side views directly with dummy data. A custom
authentication backend will authenticate any user without checking
credentials.