Fix API so speakers don't get extra spaces
NDV keeps pushing data where the speaker name has extra spaces before or
after the name which causes it not to match the equivalent speaker.
E.g. ' Carl' doesn't match 'Carl', so we end up with two Carls in the
db.

This fixes that by stripping the speaker name before the db lookup.