Change to django-configurations
There's a LOT of movement and re-writing of comments. End results, the
base settings should be the same, and the comments should link to the
current django documentation. The structure is mainly based on
django-cookiecutter, but I didn't add settings from that. I just moved
though around to line up (mostly) with it.