Remove suggestions
This was a stopgap solution to allow people to suggest other conferences
we should add. It stunk, though, since we had spam problems and we
really needed to expand this to a more full-fledged issue tracker in
order to effectively track progress on adding conferences.

At that point, it's way better to use a real issue tracker and to ditch
this app.

Fixes #277