Put database.db and whoosh_index in repo root
Previously the database.db and whoosh_index files would get created in
src/richard/ along with the source code which I think is a little weird.
I moved up a couple of directories so it's in the repository root.

It might be a better idea to put them in a data/ directory so you could
wipe the state of richard just by deleting that directory.